requirements.txt
imageio==2.30.0
matplotlib==3.7.1
numpy==1.23.5
Pillow==9.5.0
scipy==1.10.1
streamlit==1.22.0
